server/requirements.txt
Flask
PyJWT
mysql-connector-python
python-dotenv
Flask-Limiter